Change returns in playRound to a variable
Chaned the returns in playRound to a variable then test the game state
before returning the variable.  This makes it that on the click where
either the computer or player gets to 5 wins the game declares a winner.

Need to also turn off eventlisteners at this point (or maybe reomove the
buttons) and put in a reset/play again button to reset game state.